# build/public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "heal_api_3001",
		"version": "1.0.0",
		"description": "Boilerplate apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
		"/user/common/contactus/fill": {
			"post": {
				"operationId": "ContactUs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Contact Us",
				"tags": [
					"User Common Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}